+++ packages/http-server/src/request-handlers/submit-rfq.ts
@@ -3,6 +3,7 @@ import type { MessageKind } from '@tbdex/protocol'
 import type { ErrorDetail } from '@tbdex/http-client'
 
 import { Message } from '@tbdex/protocol'
+import { CallbackError } from './index.js'
 
 type SubmitRfqOpts = {
   callback: SubmitCallback<'rfq'>
@@ -54,8 +55,14 @@ export function submitRfq(options: SubmitRfqOpts): RequestHandler {
 
     try {
       // TODO: figure out what to do with callback result, if anything. (issue #7)
-      const _result = await callback({ request: req, response: res }, message)
+      const _result = await callback({ request: req, response: res }, message, { offering })
     } catch(e) {
+      if (e instanceof CallbackError) {
+        return res.status(e.statusCode).json({ errors: e.details })
+      } else {
+        const errorDetail: ErrorDetail = { detail: 'umm idk' }
+        return res.status(500).json({ errors: [errorDetail] })
+      }
       // TODO: handle error lewl (#issue 8)
     }